Lord, pluck out my eyes so I don't see this abomination
This code is ugly, no denying it.  Bootstrapping is ugly.  We're in a minimal
environment, and our tool to install the tools depends on other tools.

The nokogiri gem installation is heinous.

I am not adept enough to see a better way, and at this point I just want
a green build.